+++ tests/testthat/test_errors.R
@@ -9,8 +9,8 @@ test_that("Function suggestions for unsupported export", {
 
 test_that("Error for unsupported file types", {
     writeLines("123", con = "test.faketype")
-    expect_error(import("test.faketype"), "Unrecognized file format")
-    expect_error(export(mtcars, "mtcars.faketype"), "Unrecognized file format")
+    expect_error(import("test.faketype"), "Format not supported")
+    expect_error(export(mtcars, "mtcars.faketype"), "Format not supported")
     expect_message(get_type("faketype"), "Unrecognized file format. Try specifying with the format argument.")
     expect_equal(get_type("faketype"), "faketype")
     expect_error(get_ext("noextension"), "'file' has no extension")
@@ -19,8 +19,8 @@ test_that("Error for unsupported file types", {
 
 test_that("Error for mixed support file types", {
     expect_error(import("test.por"), "No such file")
-    expect_error(export(mtcars, "mtcars.por"), "Unrecognized file format")
-    expect_error(export(mtcars, "mtcars.faketype"), "Unrecognized file format")
+    expect_error(export(mtcars, "mtcars.por"), "Format not supported")
+    expect_error(export(mtcars, "mtcars.faketype"), "Format not supported")
 })
 
 test_that("Only export data.frame or matrix", {